Final backend commit

+++ src/main/java/com/logincontroller/filmrentalsystem/dto/AddressDTO.java
@@ -1,17 +1,18 @@
 package com.logincontroller.filmrentalsystem.dto;
 
 import lombok.Data;
+
 import java.time.LocalDateTime;
 
 @Data
 public class AddressDTO {
-    private Integer addressId;
+    private Short addressId;
     private String address;
     private String address2;
     private String district;
-    private Integer cityId;
+    private Short cityId;
     private String cityName;
-    private Integer countryId;
+    private Short countryId;
     private String countryName;
     private String postalCode;
     private String phone;
