+++ src/main/java/com/logincontroller/filmrentalsystem/dto/CountryDTO.java
@@ -1,11 +1,12 @@
 package com.logincontroller.filmrentalsystem.dto;
 
 import lombok.Data;
+
 import java.time.LocalDateTime;
 
 @Data
 public class CountryDTO {
-    private Integer countryId;
+    private Short countryId;
     private String country;
     private LocalDateTime lastUpdate;
 }
