+++ src/main/java/com/logincontroller/filmrentalsystem/dto/StoreDTO.java
@@ -1,15 +1,16 @@
 package com.logincontroller.filmrentalsystem.dto;
 
 import lombok.Data;
+
 import java.time.LocalDateTime;
 
 @Data
 public class StoreDTO {
-    private Integer storeId;
-    private Integer managerStaffId;
+    private Byte storeId;
+    private Byte managerStaffId;
     private String managerFirstName;
     private String managerLastName;
-    private Integer addressId;
+    private Short addressId;
     private String addressLine;
     private String city;
     private String country;
