+++ src/main/java/com/logincontroller/filmrentalsystem/repository/InventoryRepository.java
@@ -8,13 +8,10 @@ import java.util.List;
 
 @Repository
 public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
-    
-    // Find all physical copies of a specific movie
+
     List<Inventory> findByFilm_FilmId(Short filmId);
 
-    // Find all physical copies sitting in a specific store
-    List<Inventory> findByStore_StoreId(Integer storeId);
-    
-    // Find a specific movie at a specific store
-    List<Inventory> findByFilm_FilmIdAndStore_StoreId(Short filmId, Integer storeId);
+    List<Inventory> findByStore_StoreId(Byte storeId);
+
+    List<Inventory> findByFilm_FilmIdAndStore_StoreId(Short filmId, Byte storeId);
 }
